feat: Make the exists operator do flow typing
Alters how `if exists r.a` works such that it does not add any
constraint that the record has the field as well as the true branch
assumes that the field always exists.

Only a proof of concept as this doesn't handle more complicated predicates.
I believe that exists in the interpreter needs to do some dynamic type
checking as it would otherwise be possible to have a field exist, but be
of the wrong type.